changes to the type of d in get_val_from_d() function and in SAFE_MIN to make it safer

+++ follow_wall/src/follow_wall/SensingNode.cpp
@@ -30,7 +30,7 @@ SensingNode::SensingNode(const std::string & name)
 }
 
 
-int get_val_from_d(int d)
+int get_val_from_d(float d)
 {
   if (d < SAFE_DISTANCE - INTERVAL) {
     return LESS;
